Split ResolveErrors to its own file.
