+++ lib/resolve_errors.go
@@ -1,0 +1,84 @@
+package sous
+
+import "fmt"
+
+type (
+	// ResolveErrors collect all the errors for a resolve action into a single
+	// error to be handled elsewhere
+	ResolveErrors struct {
+		Causes []error
+	}
+
+	// MissingImageNamesError reports that we couldn't get names for one or more
+	// source versions
+	MissingImageNamesError struct {
+		Causes []error
+	}
+	// CreateError is returned when there's an error trying to create a deployment
+	CreateError struct {
+		Deployment *Deployment
+		Err        error
+	}
+
+	// DeleteError is returned when there's an error while trying to delete a deployment
+	DeleteError struct {
+		Deployment *Deployment
+		Err        error
+	}
+
+	// ChangeError describes an error that occurred while trying to change one deployment into another
+	ChangeError struct {
+		Deployments *DeploymentPair
+		Err         error
+	}
+
+	// RectificationError is an interface that extends error with methods to get
+	// the deployments the preceeded and were intended when the error occurred
+	RectificationError interface {
+		error
+		ExistingDeployment() *Deployment
+		IntendedDeployment() *Deployment
+	}
+)
+
+func (e *CreateError) Error() string {
+	return fmt.Sprintf("Couldn't create deployment %+v: %v", e.Deployment, e.Err)
+}
+
+// ExistingDeployment returns the deployment that was already existent in a change error
+func (e *CreateError) ExistingDeployment() *Deployment {
+	return nil
+}
+
+// IntendedDeployment returns the deployment that was intended in a ChangeError
+func (e *CreateError) IntendedDeployment() *Deployment {
+	return e.Deployment
+}
+
+func (e *DeleteError) Error() string {
+	return fmt.Sprintf("Couldn't delete deployment %+v: %v", e.Deployment, e.Err)
+}
+
+// ExistingDeployment returns the deployment that was already existent in a change error
+func (e *DeleteError) ExistingDeployment() *Deployment {
+	return e.Deployment
+}
+
+// IntendedDeployment returns the deployment that was intended in a ChangeError
+func (e *DeleteError) IntendedDeployment() *Deployment {
+	return nil
+}
+
+func (e *ChangeError) Error() string {
+	return fmt.Sprintf("Couldn't change from deployment %+v to deployment %+v: %v", e.Deployments.Prior, e.Deployments.Post, e.Err)
+}
+
+// ExistingDeployment returns the deployment that was already existent in a change error
+func (e *ChangeError) ExistingDeployment() *Deployment {
+	return e.Deployments.Prior
+}
+
+// IntendedDeployment returns the deployment that was intended in a ChangeError
+func (e *ChangeError) IntendedDeployment() *Deployment {
+	return e.Deployments.Post
+}
